perf(cli): defer the heavy plugin-dispatch graph off the startup path
The host CLI entry statically imported `./plugin-commands-dispatch`, whose
transitive deps construct the full @/config + @/config/providers + @/plugin +
zod + @/container module graph (~190ms). That cost was paid on every
invocation — `start`, `stop`, `tui`, `--version`, `--help` — even though
`dispatchPluginCommand` only ever runs for non-builtin (plugin) commands.

The entry also pulled `findAgentDir` from the 1590-line `../init` barrel,
which statically imports the same heavy graph (@/config, @/container,
@/secrets, @/tui). `findAgentDir` itself is a trivial fs walk needed at the
migration dispatch boundary, so the barrel forced the whole graph onto every
real host subcommand before its lazy command module even loaded.

Changes:
- Extract `findAgentDir`/`isInitialized`/`CONFIG_FILE` into a dependency-free
  `src/init/find-agent-dir.ts` (only node:fs + node:path). The barrel
  re-exports them, so existing importers and their tests are untouched.
- Import `findAgentDir` in `src/cli/index.ts` from the tiny module.
- Defer `dispatchPluginCommand` to a dynamic `import()` inside the non-builtin
  branch, after the host startup migration (preserving the documented
  migration-before-dispatch ordering; bare flags and `run` still skip it).
- Add a guard test asserting the new module imports nothing beyond node:fs and
  node:path, so a future heavy import can't silently reintroduce the cost.

Entry static graph drops ~190ms -> ~20ms; `typeclaw --version` likewise.
No behavior change: 965 existing init/cli/plugin tests pass unchanged.
